ibmcloud-powervs: add a feature to send user-data via SFTP
This PR adds a feature in ibmcloud-powervs provider to use an alternate
way to send user data to the peerpod VM via SSH File Transfer Protocol(SFTP).

Initially, we generate an SSH key value pair when the provider is initialised.
When user sets --enable-sftp to true, we pass the public key as user-data during the
VM creation. Once the VM is reachable, the provider then connects to it to
securely transfer the user data to path /media/cidata/user-data after which
process-user-data would follow its usual flow of processing it, similar to the
docker provider.

Signed-off-by: Amulyam24 <amulmek1@in.ibm.com>